requirements.txt
json5==0.12.1
python-dotenv==1.0.1
requests==2.32.3